+++ frontend/app/[locale]/components/company/AddSiteToCompanyForm.tsx
@@ -2,7 +2,7 @@
 
 import React, { useEffect, useState } from "react"
 import { SubscribeEvent } from "@/app/[locale]/components/base/SubscribeEvent"
-import { FormLayout } from "@/app/[locale]/components/layout/formLayout"
+import { FormModalLayout } from "@/app/[locale]/components/layout/formModalLayout"
 import { EventLogLayoutMin2 } from "@/app/[locale]/components/layout/EventFieldLayout"
 import {
     FormInputLayout2,
@@ -15,7 +15,6 @@ import {
     IhandleEventsResponse,
     ILayoutEventLogMapping
 } from "@/interfaces/layout"
-import { useToast } from "@chakra-ui/react"
 import {
     IGeneral,
     ILayoutButton,
@@ -25,41 +24,38 @@ import {
     generalIntl,
     layoutButtonIntl, layoutEventLogIntl, toasterMessages
 } from "@/utils/intl"
-import { createSiteFormIntl, ICreateSiteForm } from "@/utils/siteIntl"
+import { addSiteToCompanyFormIntl, IAddSiteToCompanyForm } from "@/utils/siteIntl"
 import { writeContractByFunctionName } from "@/utils"
 import { ICreateSiteModalForm } from "@/interfaces/registers"
 import { handleEvents } from "@/utils/events"
 import { useGlobalContext } from "@/contexts/Global"
 import { useIdentityContext } from "@/contexts/Identity"
+import RestrictArea from "@/app/[locale]/components/RestrictArea"
 
-const CreateSiteModalForm = ({ props }: { props: ICreateSiteModalForm }) => {
-    const componentName: string = "CreateSiteModalForm"
-
+export const AddSiteToCompanyForm = ({ props }: { props: ICreateSiteModalForm }) => {
     // contexts
     const { address } = useAccount()
     const { subscribeEventLogs, setSubscribeEventLogs } = useRegisterSecurityEventContext()
     const { setReloadPage } = useGlobalContext()
     const { company } = useIdentityContext()
+
     // states
     const [loading, setLoading] = useState(false)
     const [eventLog, setEventLog] = useState<IEventLog | undefined>()
-    const [eventName, setEventName] = useState("RegisterCreated")
     const [eventWaiting, setEventWaiting] = useState(false)
+    const [blockQuoteLevel, setBlockQuoteLevel] = useState("")
+    const [blockQuoteMessage, setBlockQuoteMessage] = useState("")
 
     // Utils
-    const toast = useToast()
-    const createSiteForm: ICreateSiteForm = createSiteFormIntl()
+    const createSiteForm: IAddSiteToCompanyForm = addSiteToCompanyFormIntl()
     const general: IGeneral = generalIntl()
     const layoutButton: ILayoutButton = layoutButtonIntl()
     const layoutEventLog: ILayoutEventLog = layoutEventLogIntl()
     const toastMessage: IToasterMessages = toasterMessages()
 
     // Constants
-    const layoutEventLogMapping: ILayoutEventLogMapping = {
-        addressAttribute: '_addr',
-        addressValue: address as `0x${string}`,
-    }
-
+    const COMPONENT_NAME: string = "CreateSiteModalForm"
+    const EVENT_NAME = 'RegisterCreated'
 
     /**
      *
@@ -76,11 +72,11 @@ const CreateSiteModalForm = ({ props }: { props: ICreateSiteModalForm }) => {
      *
      */
     const [formData, setFormData] = useState<FormDataLayout>({
-        name: company.name,
-        addressName: company.name,
-        siret: company.siret,
+        name: company.name || "",
+        addressName: company.name || "",
+        siret: company.siret || "",
         siteName: "",
-        siteAddressName: "",
+        siteAddressName: ""
     })
 
     /**
@@ -109,50 +105,45 @@ const CreateSiteModalForm = ({ props }: { props: ICreateSiteModalForm }) => {
             })
             .catch(err => {
                 console.log(`writeContractByFunctionName(${actionName}) error => ` + err)
-                toast({
-                    title: toastMessage.addNewSiteToCompanyErrorTitle,
-                    description: toastMessage.addNewSiteToCompanyErrorDescription,
-                    status: 'error',
-                    duration: 5000,
-                    isClosable: true,
-                })
+                setBlockQuoteMessage(toastMessage.addSiteToCompanyErrorDescription)
+                setBlockQuoteLevel("error")
             })
     }
 
     useEffect(() => {
-        if ( subscribeEventLogs.id === componentName && subscribeEventLogs.eventLog.length > 0 ) {
+        const layoutEventLogMapping: ILayoutEventLogMapping = {
+            addressAttribute: '_addr',
+            addressValue: address as `0x${string}`,
+        }
+
+        if ( subscribeEventLogs.id === COMPONENT_NAME && subscribeEventLogs.eventLog.length > 0 ) {
             const handleEventsResponse: IhandleEventsResponse = handleEvents(
                 layoutEventLog, subscribeEventLogs.eventLog, layoutEventLogMapping)
 
             if (handleEventsResponse.ok) {
                 setEventLog(handleEventsResponse.eventLog)
-                toast({
-                    title: toastMessage.addNewSiteToCompanyOkTitle,
-                    description: toastMessage.addNewSiteToCompanyOkDescription,
-                    status: 'success',
-                    duration: 5000,
-                    isClosable: true,
-                })
+                setBlockQuoteMessage(toastMessage.addSiteToCompanyOkDescription)
+                setBlockQuoteLevel("success")
             }
             setEventWaiting(false)
             setLoading(false)
-            setSubscribeEventLogs({id: componentName, eventLog: []})
+            setSubscribeEventLogs({id: COMPONENT_NAME, eventLog: []})
             setReloadPage(true)
         }
     }, [subscribeEventLogs])
 
-    console.log('company', JSON.stringify(company))
-
     return (
-        <>
-            {eventWaiting && <SubscribeEvent props={{taskId: componentName, eventName: eventName}} />}
+        <RestrictArea asCompany={true}>
+            {eventWaiting && <SubscribeEvent props={{taskId: COMPONENT_NAME, eventName: EVENT_NAME}} />}
 
             {props.isModalOpen &&
                 <div className="pt-0 z-10">
-                    <FormLayout props={{
+                    <FormModalLayout props={{
                         title: createSiteForm.title,
                         onModalClose: props.closeModal,
-                        onSubmit: handleSubmit
+                        onSubmit: handleSubmit,
+                        message: blockQuoteMessage,
+                        messageLevel: blockQuoteLevel
                     }}>
                         {/*Event log*/}
                         <div className="my-2">
@@ -192,11 +183,9 @@ const CreateSiteModalForm = ({ props }: { props: ICreateSiteModalForm }) => {
                                 height: 'py-2 lg:py-3'
                             }}/>
                         </div>
-                    </FormLayout>
+                    </FormModalLayout>
                 </div>
             }
-        </>
+        </RestrictArea>
     )
 }
-
-export default CreateSiteModalForm